+++ Release_Notes.md
@@ -26,6 +26,7 @@
     - DataAccess.SQLServer\DMessages.Messages
     - DataAccess.SQLServer\DMessages.Save
     - DataAccess.SQLServer\DNameValuePairs.DeleteNVPDetail
+    - DataAccess.SQLServer\DAccounts.RefreshTokenExists
 
 ## Net_Core Version 6.0.0.0 (2025-05-16)
 ### Deprecated Classes